+++ example/lib/main.dart
@@ -64,8 +64,10 @@ const kCustomPagedDataTableTheme = PagedDataTableThemeData(
     headerBackgroundColor: Color(0xFF80CBC4),
     filtersHeaderBackgroundColor: Color(0xFF80CBC4),
     footerBackgroundColor: Color(0xFF80CBC4),
-    textStyle: TextStyle(decoration: TextDecoration.underline, fontWeight: FontWeight.normal),
+    textStyle: TextStyle(
+        decoration: TextDecoration.underline, fontWeight: FontWeight.normal),
     rowsTextStyle: TextStyle(decoration: TextDecoration.overline),
+    footerButtonsColor: Colors.white,
     chipTheme: ChipThemeData(
         backgroundColor: Colors.teal,
         labelStyle: TextStyle(color: Colors.white),